Add new TripsNetworkServlet and astract TripDataConverter helpers
